# Library change for compress middleware to increase performance