### What does this PR do?

This PR changes the use of [github.com/traefik/gziphandler](https://github.com/traefik/gziphandler) (based on [https://github.com/nytimes/gziphandler](https://github.com/nytimes/gziphandler)) to [github.com/klauspost/compress/gzhttp](https://github.com/klauspost/compress/tree/master/gzhttp).

According to some simple [bench test](https://github.com/klauspost/compress/tree/master/gzhttp#performance), this lib change allows for a huge performance boost.

<details><summary>Benchmark of library change</summary>

```bash
> benchcmp bench/old.txt bench/newlib.txt
benchmark                      old ns/op     new ns/op     delta
BenchmarkCompress_S2k-12       73087         34939         -52.20%
BenchmarkCompress_S20k-12      133451        42235         -68.35%
BenchmarkCompress_S100k-12     401413        85902         -78.60%
BenchmarkCompress_P2k-12       11206         5719          -48.96%
BenchmarkCompress_P20k-12      21733         7162          -67.05%
BenchmarkCompress_P100k-12     69122         14445         -79.10%

benchmark                      old MB/s     new MB/s     speedup
BenchmarkCompress_S2k-12       28.02        58.62        2.09x
BenchmarkCompress_S20k-12      153.47       484.91       3.16x
BenchmarkCompress_S100k-12     255.10       1192.06      4.67x
BenchmarkCompress_P2k-12       182.76       358.08       1.96x
BenchmarkCompress_P20k-12      942.35       2859.41      3.03x
BenchmarkCompress_P100k-12     1481.43      7088.99      4.79x

benchmark                      old allocs     new allocs     delta
BenchmarkCompress_S2k-12       47             29             -38.30%
BenchmarkCompress_S20k-12      48             30             -37.50%
BenchmarkCompress_S100k-12     48             31             -35.42%
BenchmarkCompress_P2k-12       47             29             -38.30%
BenchmarkCompress_P20k-12      48             30             -37.50%
BenchmarkCompress_P100k-12     48             31             -35.42%

benchmark                      old bytes     new bytes     delta
BenchmarkCompress_S2k-12       7578          3367          -55.57%
BenchmarkCompress_S20k-12      31043         4828          -84.45%
BenchmarkCompress_S100k-12     132514        6874          -94.81%
BenchmarkCompress_P2k-12       6768          2887          -57.34%
BenchmarkCompress_P20k-12      27331         3990          -85.40%
BenchmarkCompress_P100k-12     116258        6047          -94.80%
```
</details>

~~I've also added in this PR a small performance improvement by creating a handler per middleware instead of per request.
It allows a much smaller performance boost in some cases.~~
 
<!--
<details><summary>Benchmark of compress middleware with my tweak</summary>

```bash
> benchcmp bench/newlib.txt bench/optim.txt
benchmark                      old ns/op     new ns/op     delta
BenchmarkCompress_S2k-12       34939         35142         +0.58%
BenchmarkCompress_S20k-12      42235         42063         -0.41%
BenchmarkCompress_S100k-12     85902         86200         +0.35%
BenchmarkCompress_P2k-12       5719          5678          -0.72%
BenchmarkCompress_P20k-12      7162          6793          -5.15%
BenchmarkCompress_P100k-12     14445         13150         -8.97%

benchmark                      old MB/s     new MB/s     speedup
BenchmarkCompress_S2k-12       58.62        58.28        0.99x
BenchmarkCompress_S20k-12      484.91       486.88       1.00x
BenchmarkCompress_S100k-12     1192.06      1187.93      1.00x
BenchmarkCompress_P2k-12       358.08       360.71       1.01x
BenchmarkCompress_P20k-12      2859.41      3015.05      1.05x
BenchmarkCompress_P100k-12     7088.99      7786.87      1.10x

benchmark                      old allocs     new allocs     delta
BenchmarkCompress_S2k-12       29             29             +0.00%
BenchmarkCompress_S20k-12      30             30             +0.00%
BenchmarkCompress_S100k-12     31             31             +0.00%
BenchmarkCompress_P2k-12       29             29             +0.00%
BenchmarkCompress_P20k-12      30             30             +0.00%
BenchmarkCompress_P100k-12     31             31             +0.00%

benchmark                      old bytes     new bytes     delta
BenchmarkCompress_S2k-12       3367          3535          +4.99%
BenchmarkCompress_S20k-12      4828          4523          -6.32%
BenchmarkCompress_S100k-12     6874          7016          +2.07%
BenchmarkCompress_P2k-12       2887          2941          +1.87%
BenchmarkCompress_P20k-12      3990          4015          +0.63%
BenchmarkCompress_P100k-12     6047          6000          -0.78%
```
</details>
-->

### Motivation

Performance optimization.

### More

- [X] Added/updated tests
- ~[ ] Added/updated documentation~

### Additional Notes

The bench was done on a "small" machine (12 intel CPU, 32 GB ram).
The PR could use a more intensive benchmark for each step.